[REFACT] Service Layer 분리
