+++ src/main/java/com/programmers/smrtstore/domain/product/application/ProductDetailService.java
@@ -1,0 +1,66 @@
+package com.programmers.smrtstore.domain.product.application;
+
+import static com.programmers.smrtstore.domain.product.application.util.ProductServiceUtil.getProduct;
+import static com.programmers.smrtstore.domain.product.application.util.ProductServiceUtil.optionValidate;
+
+import com.programmers.smrtstore.core.properties.ErrorCode;
+import com.programmers.smrtstore.domain.product.application.dto.req.CreateProductDetailOptionRequest;
+import com.programmers.smrtstore.domain.product.application.dto.req.CreateProductRequest;
+import com.programmers.smrtstore.domain.product.application.dto.req.ProductDetailOptionRequest;
+import com.programmers.smrtstore.domain.product.application.dto.res.ProductDetailOptionResponse;
+import com.programmers.smrtstore.domain.product.application.dto.res.ProductResponse;
+import com.programmers.smrtstore.domain.product.domain.entity.Product;
+import com.programmers.smrtstore.domain.product.domain.entity.ProductDetailOption;
+import com.programmers.smrtstore.domain.product.exception.ProductException;
+import com.programmers.smrtstore.domain.product.infrastructure.ProductDetailOptionJpaRepository;
+import com.programmers.smrtstore.domain.product.infrastructure.ProductJpaRepository;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class ProductDetailService {
+
+    private final ProductJpaRepository productRepository;
+    private final ProductDetailOptionJpaRepository detailOptionRepository;
+
+    // 상품 추가 다중 옵션
+    public ProductResponse createProduct(CreateProductRequest request,
+        List<CreateProductDetailOptionRequest> optionRequests) {
+        optionValidate(request.isCombinationYn());
+        Product product = request.toEntity();
+        List<ProductDetailOption> productOptions = optionRequests.stream()
+            .map(optionRequest -> optionRequest.toEntity(product))
+            .toList();
+        Product result = productRepository.save(product);
+        detailOptionRepository.saveAll(productOptions);
+        return ProductResponse.from(result);
+    }
+
+    public ProductDetailOptionResponse addProductDetailOption(Long productId,
+        CreateProductDetailOptionRequest request) {
+        Product product = getProduct(productRepository, productId);
+        var detailOption = detailOptionRepository.save(request.toEntity(product));
+        return ProductDetailOptionResponse.from(detailOption);
+    }
+
+    public Long removeProductOption(Long productId, Long productOptionId) {
+        Product product = getProduct(productRepository, productId);
+        optionValidate(product.isCombinationYn());
+        product.removeOption(productOptionId);
+        return productOptionId;
+    }
+
+    public ProductDetailOptionResponse updateProductOption(
+        ProductDetailOptionRequest request) {
+        ProductDetailOption option = detailOptionRepository.findById(request.getId())
+            .orElseThrow(() -> new ProductException(
+                ErrorCode.PRODUCT_OPTION_NOT_FOUND));
+        option.updateValues(request.getQuantity(), request.getPrice(), request.getOptionNames());
+        return ProductDetailOptionResponse.from(option);
+    }
+
+}
